requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-filter==21.1
django-js-asset==1.2.2
django-nested-admin==3.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.3
python-monkey-business==1.0.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.1
whitenoise==5.3.0
